pyproject.toml
[project]
name = "mmsegmentation"
version = "1.2.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "codecov>=2.1.13",
    "diffusers>=0.35.2",
    "einops==0.3.0",
    "flake8>=7.3.0",
    "ftfy>=6.3.1",
    "imageio==2.9.0",
    "imageio-ffmpeg==0.4.2",
    "interrogate>=1.7.0",
    "invisible-watermark>=0.2.0",
    "kornia==0.6",
    "matplotlib>=3.10.7",
    "mmcv==2.1.0",
    "mmdet>=3.1.0",
    "mmengine>=0.10.7",
    "nibabel>=5.3.2",
    "ninja>=1.13.0",
    "numpy>=2.3.4",
    "omegaconf==2.1.1",
    "opencv-python>=4.11.0.86",
    "packaging>=25.0",
    "pillow>=11.3.0",
    "prettytable>=3.16.0",
    "psutil>=7.1.1",
    "pudb==2019.2",
    "pytest>=8.4.2",
    "pytorch-lightning==1.4.2",
    "regex>=2025.9.18",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "streamlit>=0.73.1",
    "test-tube>=0.7.5",
    "timm>=1.0.20",
    "torch-fidelity==0.3.0",
    "torchmetrics==0.6.0",
    "tqdm>=4.67.1",
    "transformers>=4.19.2",
    "xdoctest>=0.10.0",
    "yapf>=0.43.0",
]

# [tool.uv.sources]
# openai-agents = { path = "/data/an/mmcv", editable = true }
